fix(benchmark): fix spector-bench compilation errors from config refactoring
