+++ bench/spector-bench/src/main/java/com/spectrayan/spector/bench/RealEmbeddingBench.java
@@ -160,7 +160,7 @@ public class RealEmbeddingBench {
                         .shardThreshold(20_000)
                         .oversamplingFactor(4)
                         .similarityFunction(SimilarityFunction.COSINE)
-                        .HnswProperties(new HnswProperties(16, 128, 64))
+                        .hnswProperties(new HnswProperties(16, 128, 64))
                         .build();
 
                 // Train on first 5000 vectors
