+++ bench/spector-bench/src/main/java/com/spectrayan/spector/bench/SpectorIndexBenchmark.java
@@ -87,7 +87,7 @@ public class SpectorIndexBenchmark {
                 .shardThreshold(20_000)
                 .oversamplingFactor(3)
                 .similarityFunction(SimilarityFunction.COSINE)
-                .HnswProperties(new HnswProperties(16, 128, 64))
+                .hnswProperties(new HnswProperties(16, 128, 64))
                 .build();
 
         // Train on a sample
