+++ bench/spector-bench/src/main/java/com/spectrayan/spector/bench/SpectorIndexQuickBench.java
@@ -68,7 +68,7 @@ public class SpectorIndexQuickBench {
                 .shardThreshold(20_000)
                 .oversamplingFactor(3)
                 .similarityFunction(SimilarityFunction.COSINE)
-                .HnswProperties(new HnswProperties(16, 128, 64))
+                .hnswProperties(new HnswProperties(16, 128, 64))
                 .build();
 
         // Train
@@ -116,7 +116,7 @@ public class SpectorIndexQuickBench {
                     .shardThreshold(20_000)
                     .oversamplingFactor(3)
                     .similarityFunction(SimilarityFunction.COSINE)
-                    .HnswProperties(new HnswProperties(16, 128, 64))
+                    .hnswProperties(new HnswProperties(16, 128, 64))
                     .build();
 
             probeIndex.train(trainVectors);
