+++ bench/spector-bench/src/main/java/com/spectrayan/spector/bench/SpectorTinyDiag.java
@@ -69,7 +69,7 @@ public class SpectorTinyDiag {
                 .shardThreshold(20_000)
                 .oversamplingFactor(10)
                 .similarityFunction(SimilarityFunction.COSINE) // user chose cosine
-                .HnswProperties(new HnswProperties(16, 128, 64))
+                .hnswProperties(new HnswProperties(16, 128, 64))
                 .build();
 
         index.train(vectors);
