+++ bench/spector-bench/src/main/java/com/spectrayan/spector/bench/cognitive/CognitiveRetriever.java
@@ -24,7 +24,7 @@ import com.spectrayan.spector.memory.model.CognitiveResult;
 import com.spectrayan.spector.memory.model.RecallMode;
 import com.spectrayan.spector.memory.model.RecallOptions;
 import com.spectrayan.spector.memory.model.ScoringMode;
-import com.spectrayan.spector.memory.model.TextSearchMode;
+import com.spectrayan.spector.config.model.TextSearchMode;
 import com.spectrayan.spector.memory.SpectorMemory;
 
 /**
