+++ bench/spector-bench/src/main/java/com/spectrayan/spector/bench/cognitive/RetrievalStackMatrixRunner.java
@@ -33,7 +33,7 @@ import com.spectrayan.spector.provider.embedding.EmbeddingProvider;
 import com.spectrayan.spector.provider.ollama.OllamaEmbeddingProvider;
 import com.spectrayan.spector.memory.model.RecallMode;
 import com.spectrayan.spector.memory.model.RecallOptions;
-import com.spectrayan.spector.memory.model.TextSearchMode;
+import com.spectrayan.spector.config.model.TextSearchMode;
 import com.spectrayan.spector.memory.SpectorMemory;
 
 /**
